Move validation util to dedicated api package
