+++ api-client/src/main/java/com/takipi/api/client/request/ViewTimeframeRequest.java
@@ -2,7 +2,7 @@ package com.takipi.api.client.request;
 
 import java.util.Collection;
 
-import com.takipi.common.util.ValidationUtil;
+import com.takipi.api.client.util.validation.ValidationUtil;
 
 public abstract class ViewTimeframeRequest extends TimeframeRequest {
 	public final String viewId;
