+++ api-client/src/main/java/com/takipi/api/client/request/event/EventMarkResolvedRequest.java
@@ -1,7 +1,7 @@
 package com.takipi.api.client.request.event;
 
 import com.takipi.api.client.request.label.ModifyLabelsRequest;
-import com.takipi.common.util.ValidationUtil;
+import com.takipi.api.client.util.validation.ValidationUtil;
 
 public class EventMarkResolvedRequest extends ModifyLabelsRequest {
 	private final String eventId;
